Add: local-publish-test
