+++ crate-test/export-test-lib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "export-test-lib"
+version = "0.1.2"
+edition = "2021"
+publish = ["test-local-registry"]
+
+[dependencies]
+
+[features]
+aaa = []
