+++ crate-test/import-test-crate/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 try-partialord = { version = "0.1.3", registry = "test-local-registry" }
 # try-partialord = { version = "0.1.3" }
+# export-test-lib = { version = "0.1.0", registry = "test-local-registry" }
