+++ crate-test/test.env
@@ -1,5 +1,5 @@
 RUST_LOG = "info,axum_cargo_registry=trace"
-REGISTRY_PATH = "../crate-test/local-storage"
+REGISTRY_PATH = "crate-test/local-storage"
 INDEX_BUCKET = "index"
 CRATE_BUCKET = "crates"
 PRESIGNED_EXPIRE = "1800"
@@ -10,3 +10,4 @@ AWS_ENDPOINT_URL = "http://localhost:9000"
 MINIO_ROOT_USER = "ROOTUSER"
 MINIO_ROOT_PASSWORD = "TESTTEST"
 MINIO_DOMAIN = localhost
+CARGO_REGISTRIES_TEST_LOCAL_REGISTRY_TOKEN=test
