+++ examples/local-axum-cargo-registry/Cargo.toml
@@ -11,3 +11,6 @@ axum = { version = "0.7.5", features = ["json", "tokio", "http1", "http2"] }
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 tokio = { version = "1", features = ["rt", "macros", "rt-multi-thread"] }
 tracing = { version = "0.1" }
+
+[features]
+api = ["axum-cargo-registry/api"]
